requirements.txt
Flask==2.0.3
Flask-JWT-Extended==4.4.4
requests==2.26.0
yfinance==0.2.51
pyjwt==2.1.0